3rd Commit:
Added - 1.get products method, 2. Integration testing using
testContainers platform

+++ product-service/src/main/java/com/explore/micro/model/Products.java
@@ -21,5 +21,5 @@ public class Products {
 	private Integer id;
 	private String name;
 	private String description;
-	private long price;
+	private Double price;
 }
